[IMP] registries: remove registries index
We had a `index.ts` file for some registries. The file was a bit of a
problem because:

- It mixes apples and oranges: some registries are purely UI, some not,
    so importing everything at once don't really make sense.
- It is an easy source of circular dependencies. For example, we had
    to split `side_panel_registry` in two files to avoid circular dependencies

This commit removes the `index.ts` file and imports the registries
individually where needed.

Task: 4737879
