--- src/registries/index.ts
@@ -1,12 +1,0 @@
-export * from "./auto_completes/index";
-export * from "./autofill_modifiers";
-export * from "./autofill_rules";
-export * from "./cell_popovers_registry";
-export * from "./chart_types";
-export * from "./currencies_registry";
-export * from "./figure_registry";
-export * from "./inverse_command_registry";
-export * from "./menus/index";
-export * from "./ot_registry";
-export * from "./side_panel_registry_entries";
-export * from "./topbar_component_registry";
