+++ tests/top_bar_component.test.ts
@@ -6,7 +6,8 @@ import { TopBar } from "../src/components/top_bar/top_bar";
 import { topBarToolBarRegistry } from "../src/components/top_bar/top_bar_tools_registry";
 import { DEBOUNCE_TIME, DEFAULT_FONT_SIZE } from "../src/constants";
 import { toZone, zoneToXc } from "../src/helpers";
-import { topbarComponentRegistry, topbarMenuRegistry } from "../src/registries";
+import { topbarMenuRegistry } from "../src/registries/menus";
+import { topbarComponentRegistry } from "../src/registries/topbar_component_registry";
 import { ConditionalFormat, Currency, Pixel, SpreadsheetChildEnv, Style } from "../src/types";
 import { FileStore } from "./__mocks__/mock_file_store";
 import { MockTransportService } from "./__mocks__/transport_service";
